Q4_work/Openai_agents_sdk_practice/recap_concepts/pyproject.toml
[project]
name = "recap-concepts"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "openai-agents>=0.1.0",
    "python-dotenv>=1.1.1",
]
